update org card to use list item, rename org card to org item
