+++ src/shared/components/Orgs/Orgs.less
@@ -3,7 +3,7 @@
 /**
 * Organisation Cards
 */
-.OrgCard {
+.OrgItem {
   background-color: #fff !important;
   margin: 5px !important;
   box-shadow: 0 2px 5px 0 rgba(49, 49, 93, 0.1), 0 1px 2px 0 rgba(0, 0, 0, 0.08);
